requirements.txt
pygame==1.9.6
matplotlib==3.2.1
imutils==0.5.3
tensorflow==2.2.0
opencv_python==4.2.0.34
numpy==1.18.5
jnius==1.1.0
Pillow==8.0.0
PySimpleGUI==4.30.0
scikit_learn==0.23.2
